backend/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
sqlalchemy
python-dotenv

